+++ src/main/java/legend/game/inventory/screens/MessageBoxScreen.java
@@ -47,7 +47,7 @@ public class MessageBoxScreen extends MenuScreen {
   }
 
   @Override
-  protected InputPropagation mouseMove(final int x, final int y) {
+  protected InputPropagation mouseMove(final double x, final double y) {
     if(super.mouseMove(x, y) == InputPropagation.HANDLED) {
       return InputPropagation.HANDLED;
     }
@@ -60,14 +60,14 @@ public class MessageBoxScreen extends MenuScreen {
     if(this.messageBox.type_15 == 2) {
       final int selectionY = this.messageBox.y_1e + 21 + this.messageBox.text_00.length * 12 / 2 - (this.messageBox.text_00.length - 1) * 3;
 
-      if(this.messageBox.menuIndex_18 != 0 && MathHelper.inBox(x, y, this.messageBox.x_1c + 4, selectionY, 112, 14)) {
+      if(this.messageBox.menuIndex_18 != 0 && MathHelper.inBox((int)x, (int)y, this.messageBox.x_1c + 4, selectionY, 112, 14)) {
         playMenuSound(1);
         this.messageBox.menuIndex_18 = 0;
 
         if(this.messageBox.highlightRenderable_04 != null) {
           this.messageBox.highlightRenderable_04.setY(selectionY - 2);
         }
-      } else if(this.messageBox.menuIndex_18 != 1 && MathHelper.inBox(x, y, this.messageBox.x_1c + 4, selectionY + 14, 112, 14)) {
+      } else if(this.messageBox.menuIndex_18 != 1 && MathHelper.inBox((int)x, (int)y, this.messageBox.x_1c + 4, selectionY + 14, 112, 14)) {
         playMenuSound(1);
         this.messageBox.menuIndex_18 = 1;
 
@@ -81,7 +81,7 @@ public class MessageBoxScreen extends MenuScreen {
   }
 
   @Override
-  protected InputPropagation mouseClick(final int x, final int y, final int button, final Set<InputMod> mods) {
+  protected InputPropagation mouseClick(final double x, final double y, final int button, final Set<InputMod> mods) {
     if(super.mouseClick(x, y, button, mods) == InputPropagation.HANDLED) {
       return InputPropagation.HANDLED;
     }
@@ -98,7 +98,7 @@ public class MessageBoxScreen extends MenuScreen {
       // Yes/no
       final int selectionY = this.messageBox.y_1e + 21 + this.messageBox.text_00.length * 12 / 2 - (this.messageBox.text_00.length - 1) * 3;
 
-      if(MathHelper.inBox(x, y, this.messageBox.x_1c + 4, selectionY, 112, 14)) {
+      if(MathHelper.inBox((int)x, (int)y, this.messageBox.x_1c + 4, selectionY, 112, 14)) {
         playMenuSound(2);
         this.messageBox.menuIndex_18 = 0;
 
@@ -108,7 +108,7 @@ public class MessageBoxScreen extends MenuScreen {
 
         this.result = MessageBoxResult.YES;
         this.messageBox.state_0c = 4;
-      } else if(MathHelper.inBox(x, y, this.messageBox.x_1c + 4, selectionY + 14, 112, 14)) {
+      } else if(MathHelper.inBox((int)x, (int)y, this.messageBox.x_1c + 4, selectionY + 14, 112, 14)) {
         playMenuSound(2);
         this.messageBox.menuIndex_18 = 1;
