+++ src/main/java/legend/game/inventory/screens/ModsScreen.java
@@ -51,7 +51,7 @@ public class ModsScreen extends VerticalLayoutScreen {
         final Label help = label.addControl(new Label("?"));
         help.setScale(0.4f);
         help.setPos((int)(help.getFont().textWidth(label.getText()) * label.getScale()) + 2, 1);
-        help.onHoverIn(() -> this.getStack().pushScreen(new TooltipScreen(I18n.translate("lod_core.ui.mods.required"), this.mouseX, this.mouseY)));
+        help.onHoverIn(() -> this.getStack().pushScreen(new TooltipScreen(I18n.translate("lod_core.ui.mods.required"), (int)this.mouseX, (int)this.mouseY)));
         this.helpLabels.put(label, help);
         this.helpText.put(label, required);
       }
